google_colab/requirements.txt
# NOAA APT Receiver - Зависимости
# Установка: pip install -r requirements.txt

# Основные зависимости
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
PyYAML>=6.0

# Визуализация
matplotlib>=3.7.0

# RTL-SDR (опционально, для работы с реальным устройством)
pyrtlsdr>=0.3.0

# Расчёт орбит спутников (опционально)
skyfield>=1.46
skyfield-data>=5

# Улучшенная обработка изображений (опционально)
opencv-python>=4.7.0

# Telegram бот (использует встроенный urllib, дополнительные зависимости не нужны)
# aiohttp>=3.8.0  # Для асинхронного Telegram бота

# Тестирование (опционально)
# pytest>=7.3.0
# pytest-cov>=4.0.0

# Линтинг (опционально)
# flake8>=6.0.0
# black>=23.0.0
